+++ src/CoinbaseTx.php
@@ -36,9 +36,11 @@ final readonly class CoinbaseTx
 
     public function totalOutputAmount(): int
     {
-        return array_sum(array_map(
-            static fn (Output $o): int => $o->amount,
-            $this->outputs,
-        ));
+        $total = 0;
+        foreach ($this->outputs as $output) {
+            $total += $output->amount;
+        }
+
+        return $total;
     }
 }